+++ src/main/java/pl/futurecollars/invoicing/db/file/FileBasedDatabase.java
@@ -20,7 +20,7 @@ public class FileBasedDatabase implements Database {
     private final JsonService jsonService;
 
     @Override
-    public int save(Invoice invoice) {
+    public long save(Invoice invoice) {
         invoice.setId(idService.getNextIdAndIncrement());
         try {
 
@@ -33,11 +33,11 @@ public class FileBasedDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> getById(int id) {
+    public Optional<Invoice> getById(long id) {
         try {
             return filesService.readAllLines(databasePath)
                 .stream()
-                .filter(line -> containsId(line, id))
+                .filter(line -> containsId(line, (int) id))
                 .map(line -> jsonService.toJavaObject(line, Invoice.class))
                 .findFirst();
         } catch (IOException ex) {
@@ -61,12 +61,12 @@ public class FileBasedDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
         try {
             List<String> allInvoices = filesService.readAllLines(databasePath);
             var listWithoutInvoiceWithGivenId = allInvoices
                 .stream()
-                .filter(line -> !containsId(line, id))
+                .filter(line -> !containsId(line, (int) id))
                 .collect(Collectors.toList());
 
             updatedInvoice.setId(id);
@@ -87,13 +87,13 @@ public class FileBasedDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> delete(int id) {
+    public Optional<Invoice> delete(long id) {
         try {
             var allInvoices = filesService.readAllLines(databasePath);
 
             var updatedList = allInvoices
                 .stream()
-                .filter(line -> !containsId(line, id))
+                .filter(line -> !containsId(line, (int) id))
                 .collect(Collectors.toList());
 
             filesService.writeLinesToFile(databasePath, updatedList);
