+++ src/main/java/pl/futurecollars/invoicing/db/jpa/JpaDatabase.java
@@ -15,13 +15,13 @@ public class JpaDatabase implements Database {
     }
 
     @Override
-    public int save(Invoice invoice) {
+    public long save(Invoice invoice) {
         return invoiceRepository.save(invoice).getId();
     }
 
     @Override
-    public Optional<Invoice> getById(int id) {
-        return invoiceRepository.findById(id);
+    public Optional<Invoice> getById(long id) {
+        return invoiceRepository.findById(Math.toIntExact(id));
     }
 
     @Override
@@ -31,7 +31,7 @@ public class JpaDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
         Optional<Invoice> invoiceOptional = getById(id);
         if (invoiceOptional.isPresent()) {
             Invoice invoice = invoiceOptional.get();
@@ -47,7 +47,7 @@ public class JpaDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> delete(int id) {
+    public Optional<Invoice> delete(long id) {
         Optional<Invoice> invoice = getById(id);
         invoice.ifPresent(invoiceRepository::delete);
         return invoice;
